requirements.txt
docling>=2.0.0
gradio>=6.0.0
lancedb>=0.16.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0
ollama>=0.3.0
pydantic>=2.8.0
pyyaml>=6.0.0
